requirements.txt
distinctipy
nltk
pandas
python-docx
python-dotenv
sentence-transformers
wordcloud
redis
requests
